+++ airflow/models/param.py
@@ -19,4 +19,6 @@
 
 from __future__ import annotations
 
-from airflow.sdk.definitions.param import *  # noqa: F403
+from airflow.sdk.definitions.param import Param, ParamsDict
+
+__all__ = ["Param", "ParamsDict"]
