+++ providers/edge/src/airflow/providers/edge/example_dags/win_notepad.py
@@ -34,7 +34,7 @@ from typing import TYPE_CHECKING, Any
 
 from airflow.models import BaseOperator
 from airflow.models.dag import DAG
-from airflow.sdk.definitions.param import Param
+from airflow.sdk import Param
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
