+++ providers/edge/src/airflow/providers/edge/example_dags/win_test.py
@@ -39,7 +39,7 @@ from airflow.models import BaseOperator
 from airflow.models.dag import DAG
 from airflow.models.variable import Variable
 from airflow.operators.empty import EmptyOperator
-from airflow.sdk.definitions.param import Param
+from airflow.sdk import Param
 from airflow.utils.operator_helpers import context_to_airflow_vars
 from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.types import ArgNotSet
